fix: close out Phase 2 verification — clippy clean, real bugs found and fixed
Ran the full Phase 2 checklist (build/test/clippy/fmt/geiger/wasm-size/coverage)
for real instead of asserting it, per QUALITY_UPGRADE_ROADMAP.md. Results:

- cargo clippy --workspace --all-targets --all-features -D warnings: fixed 33
  dead-code/unused-import errors plus assorted unrelated lint issues (duplicated
  cfg(test) attributes, unnecessary casts, deprecated register_contract, a
  copy-pasted duplicate line, const-foldable asserts). Now clean.
- cargo fmt --all -- --check: now clean.
- cargo test --workspace: found and fixed one real pre-existing failure — a
  version() fn was added to trustforge_delegation's public interface without
  refreshing the pinned contractspecv0 XDR snapshot (CONTRACT_SPEC_VERSION 1->2).
- Discovered ~60 files (~18,500 lines) in trustforge_bond that were never
  declared as `mod` in lib.rs and so were never compiled — real feature modules
  (verifier staking, governance-voted slashing, evidence, access control,
  cooldowns, fees, status snapshots) plus most of the crate's tests. Documented
  in docs/ORPHANED_MODULES.md rather than mechanically restored: a restoration
  attempt surfaced ~45 missing #[contractimpl] entrypoints and 1,279 compile
  errors, which is new-ABI-surface design work, not a lint fix.
- Two of the dead-code fixes flagged real gaps rather than being silently
  allowed away: normalization.rs's decimal scaling and pausable.rs's pause
  enforcement are implemented but never invoked by any production entrypoint.
- trustforge_bond.wasm is 137KB against Soroban's ~64KB deploy limit, even with
  full size optimization already applied — not currently deployable anywhere.
- Coverage (cargo llvm-cov, matching what coverage.yml actually runs): timelock
  95.80%, trustforge_delegation 95.85%, trustforge_bond 70.64% (below CI's 95%
  gate; upgrade_auth.rs is at a flat 0.00%).
- Two 10,000-case proptest files (proptest_tier.rs, nonce_replay.rs) make a
  literal cargo test --workspace run take ~2+ hours — likely why the roadmap's
  baseline couldn't confirm tests pass in CI. Not a correctness bug; flagged as
  a CI-runtime problem needing a deliberate case-count decision.

Full log in VERIFICATION.md. QUALITY_UPGRADE_ROADMAP.md and STATUS.md updated
to reflect all of the above.

+++ contracts/trustforge_registry/src/lib.rs
@@ -20,10 +20,10 @@
 //! - Validates addresses before registration
 //! - Emits events for audit trail
 
-use soroban_sdk::String;
-use trustforge_errors::ContractError;
 use soroban_sdk::panic_with_error;
+use soroban_sdk::String;
 use soroban_sdk::{contract, contractimpl, contracttype, Address, Env, Symbol, Vec};
+use trustforge_errors::ContractError;
 pub mod idempotency;
 
 const STORAGE_TTL_EXTEND_TO: u32 = 31_536_000;
@@ -460,7 +460,7 @@ impl TrustForgeRegistry {
             .unwrap_or_else(|| Vec::new(&e));
 
         let actual_limit = limit.min(MAX_IDENTITIES_PAGE_SIZE);
-        let total_count = all_identities.len() as u32;
+        let total_count = all_identities.len();
 
         // Handle offset past end
         if offset >= total_count {
@@ -472,7 +472,7 @@ impl TrustForgeRegistry {
 
         let mut result = Vec::new(&e);
         for i in start..end {
-            result.push_back(all_identities.get(i as u32).unwrap());
+            result.push_back(all_identities.get(i).unwrap());
         }
 
         result
@@ -788,4 +788,3 @@ fn constant_time_eq(left: &[u8], right: &[u8]) -> bool {
         .fold(0, |acc, (l, r)| acc | (l ^ r))
         == 0
 }
-
